char counter works but maybe i have to consider stop adding once it has failed, because i think so far if the loop is not fully over and a write failes it would just count from -1 back up again

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/08 21:27:17 by jonask            #+#    #+#             */
-/*   Updated: 2023/10/18 10:27:19 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/10/20 09:40:53 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,21 +16,21 @@
 # include <sys/_types/_va_list.h>
 
 //helper.c
-int		is_specifier(char identifier, char type);
-void	print_argument(va_list *to_print, int type, int	*writecount);
-int		count_args(const char *str);
-long	reverse_num(unsigned long num);
+int			is_specifier(char identifier, char type);
+void		print_argument(va_list *to_print, int type, int	*writecount);
+int			count_args(const char *str);
+long long	reverse_num(unsigned long long num);
 
 //conversion.c
-void	put_string(char *str, int *writecount);
-void	put_char(char c, int *writecount);
-void	put_number(int num, int *writecount);
-void	put_long(long num, int *writecount);
-void	put_double(double num, int *writecount);
-void	put_float(float num, int *writecount);
-void	put_pointer(void *ptr, int *writecount);
+void		put_string(char *str, int *writecount);
+void		put_char(char c, int *writecount);
+void		put_number(int num, int *writecount);
+void		put_long(long num, int *writecount);
+void		put_double(double num, int *writecount);
+void		put_float(float num, int *writecount);
+void		put_pointer(void *ptr, int *writecount);
 
 //base_system_conversion.c
-void	put_base(long num, int base, int capital, int *writecount);
+void		put_base(long num, int base, int capital, int *writecount);
 
 #endif
